third_party/blink/web_tests/editing/execCommand/delete-textarea-readonly-removed.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<textarea id="ta" readonly>FAIL</textarea>
<script>
  test(() => {
    ta.select();
    ta.removeAttribute("readonly");
    document.execCommand("delete");
    assert_equals(ta.value, "", "The textarea.value should be empty after delete.");
  }, "Delete textarea text with execCommand after removing readonly attribute.");
</script>
